Added better charts model

+++ Charts/Charts/Charts/ViewController.swift
@@ -35,7 +35,7 @@ class ViewController: UIViewController {
     presenter.viewWillAppear()
   }
 
-  func updateValues(_ values: [Int], max: Int) {
+  func updateValues(_ values: [Double], max: Double) {
     contentView.updateValues(values, max: max)
   }
 }
